Re-implementation of #59 (a rebase and cleanup of #88) (#100)
Re-implementation of #59 (a rebase and cleanup of #88)

Adds a single new method lang.

This methods allows creating a shallow copy of the
FluentLanguageLoader which can than be used just like the original
loader but with a different current language setting. That makes it
possible to use the fl! macro without any changes and is a far more
elegant implementation than adding multiple get_lang* methods as
done in #84.

Co-authored-by: Michael P. Jung <michael.jung@terreon.de>