i18n-embed-fl: check args on MessageReference (#146)
* i18n-embed: retrieve fluent bundle with message

Add `FluentLanguageLoader::with_fluent_message_and_bundle()`. This
method functions like `with_fluent_message()`, but it also returns
the `FluentBundle` which owns the message.

This addition permits the closure to resolve `MessageReference`
and to retrieve other messages on the same bundle.

* fix(i18n-embed-fl): check args on MessageReference

The `fl!()` macro checks that all assigned arguments exist on
the message in question. But fluent messages can also refer to
other messages, and those messages can also have arguments.

```ftl
hello = Hello { to-you }
to-you = to you, {$name}!
```

Previous versions of the macro could not check arguments on
`MessageReference`s since it had no way to resolve them.

Bring the owning `FluentBundle` into scope so that message
references can be resolved. Some unnecessary generics on the
recursive functions are made concrete.

This patch requires an update to i18-embed for new API.